requirements.txt
websockets
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
importlib-metadata==0.17
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
psutil==5.6.2
py==1.8.0
pydantic==0.27
pylint==2.3.1
pyparsing==2.4.0
pytest==4.6.1
pytest-mock==1.10.4
six==1.12.0
typed-ast==1.3.5
wcwidth==0.1.7
wrapt==1.11.1
zipp==0.5.1
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
importlib-metadata==0.17
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
psutil==5.6.2
py==1.8.0
pylint==2.3.1
pyparsing==2.4.0
pytest==4.6.1
six==1.12.0
typed-ast==1.3.5
wcwidth==0.1.7
wrapt==1.11.1
zipp==0.5.1
numpy==1.17.2
requests==2.22.0
amplitude-python==0.13